Split validation file
Validation.kt was very large. It is now split into multiple files for
better understandability.
